+++ CHANGELOG.md
@@ -16,6 +16,7 @@
  - IMDb: 
    - Fix runtime scraping
    - Fix rating scraping of TV episodes (#1441)
+ - HotMovies: Search and scraping of ratings was fixed
 
 ### Changes